Move youth profile related exceptions into youths app
Refs OM-428
