+++ open_city_profile/consts.py
@@ -18,15 +18,3 @@ CONNECTED_SERVICE_DELETION_NOT_ALLOWED_ERROR = (
 INVALID_EMAIL_FORMAT_ERROR = "INVALID_EMAIL_FORMAT"
 PROFILE_MUST_HAVE_ONE_PRIMARY_EMAIL = "PROFILE_MUST_HAVE_ONE_PRIMARY_EMAIL"
 SERVICE_CONNECTION_ALREADY_EXISTS_ERROR = "SERVICE_CONNECTION_ALREADY_EXISTS_ERROR"
-
-# Youth Profile specific errors
-APPROVER_EMAIL_CANNOT_BE_EMPTY_FOR_MINORS_ERROR = (
-    "APPROVER_EMAIL_CANNOT_BE_EMPTY_FOR_MINORS_ERROR"
-)
-CANNOT_CREATE_YOUTH_PROFILE_IF_UNDER_13_YEARS_OLD_ERROR = (
-    "CANNOT_CREATE_YOUTH_PROFILE_IF_UNDER_13_YEARS_OLD_ERROR"
-)
-CANNOT_RENEW_YOUTH_PROFILE_ERROR = "CANNOT_RENEW_YOUTH_PROFILE_ERROR"
-CANNOT_SET_PHOTO_USAGE_PERMISSION_IF_UNDER_15_YEARS_ERROR = (
-    "CANNOT_SET_PHOTO_USAGE_PERMISSION_IF_UNDER_15_YEARS_ERROR"
-)
