+++ youths/consts.py
@@ -1,0 +1,11 @@
+# Youth Profile specific errors
+APPROVER_EMAIL_CANNOT_BE_EMPTY_FOR_MINORS_ERROR = (
+    "APPROVER_EMAIL_CANNOT_BE_EMPTY_FOR_MINORS_ERROR"
+)
+CANNOT_CREATE_YOUTH_PROFILE_IF_UNDER_13_YEARS_OLD_ERROR = (
+    "CANNOT_CREATE_YOUTH_PROFILE_IF_UNDER_13_YEARS_OLD_ERROR"
+)
+CANNOT_RENEW_YOUTH_PROFILE_ERROR = "CANNOT_RENEW_YOUTH_PROFILE_ERROR"
+CANNOT_SET_PHOTO_USAGE_PERMISSION_IF_UNDER_15_YEARS_ERROR = (
+    "CANNOT_SET_PHOTO_USAGE_PERMISSION_IF_UNDER_15_YEARS_ERROR"
+)
