+++ youths/tests/test_youth_profiles_graphql_api.py
@@ -8,11 +8,7 @@ from graphql_relay.node.node import to_global_id
 from guardian.shortcuts import assign_perm
 
 from open_city_profile.consts import (
-    APPROVER_EMAIL_CANNOT_BE_EMPTY_FOR_MINORS_ERROR,
-    CANNOT_CREATE_YOUTH_PROFILE_IF_UNDER_13_YEARS_OLD_ERROR,
     CANNOT_PERFORM_THIS_ACTION_WITH_GIVEN_SERVICE_TYPE_ERROR,
-    CANNOT_RENEW_YOUTH_PROFILE_ERROR,
-    CANNOT_SET_PHOTO_USAGE_PERMISSION_IF_UNDER_15_YEARS_ERROR,
     PERMISSION_DENIED_ERROR,
 )
 from open_city_profile.tests.factories import GroupFactory
@@ -20,6 +16,12 @@ from profiles.enums import EmailType
 from profiles.models import Profile
 from profiles.tests.factories import EmailFactory, ProfileWithPrimaryEmailFactory
 from services.enums import ServiceType
+from youths.consts import (
+    APPROVER_EMAIL_CANNOT_BE_EMPTY_FOR_MINORS_ERROR,
+    CANNOT_CREATE_YOUTH_PROFILE_IF_UNDER_13_YEARS_OLD_ERROR,
+    CANNOT_RENEW_YOUTH_PROFILE_ERROR,
+    CANNOT_SET_PHOTO_USAGE_PERMISSION_IF_UNDER_15_YEARS_ERROR,
+)
 from youths.enums import YouthLanguage
 from youths.tests.factories import ProfileFactory, YouthProfileFactory
